Rename notebooks folder and move some functions to utils file
